orchestrator/backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
pydantic>=2.0.0
snowflake-connector-python>=3.6.0
langgraph>=1.0.0
langchain-core>=0.3.0
langfuse>=2.0.0
httpx>=0.28.0
